+++ src/ui/svelte/task-switcher/index.ts
@@ -1,4 +1,3 @@
-import { disableAnimationsOnPerformanceMode } from "libs/ui/react/utils/performance";
 import { mount } from "svelte";
 import App from "./App.svelte";
 import { Widget } from "@seelen-ui/lib";
@@ -11,8 +10,6 @@ const root = document.getElementById("root")!;
 const widget = Widget.getCurrent();
 await widget.init();
 await widget.webview.setResizable(false);
-
-disableAnimationsOnPerformanceMode();
 
 mount(App, {
   target: root,