+++ src/ui/svelte/wallpaper_manager/index.ts
@@ -1,6 +1,5 @@
 import { mount } from "svelte";
 import { Widget } from "@seelen-ui/lib";
-import { disableAnimationsOnPerformanceMode } from "libs/ui/react/utils/performance.ts";
 import App from "./app.svelte";
 import { loadTranslations } from "./i18n/index.ts";
 
@@ -11,7 +10,6 @@ await Widget.getCurrent().init();
 Widget.self.webview.setFocusable(false);
 
 await loadTranslations();
-disableAnimationsOnPerformanceMode();
 
 const root = document.getElementById("root")!;
 mount(App, { target: root });