+++ src/ui/svelte/window_manager/index.ts
@@ -2,7 +2,6 @@ import { mount } from "svelte";
 import { Widget } from "@seelen-ui/lib";
 import { getRootContainer } from "libs/ui/react/utils";
 import { declareDocumentAsLayeredHitbox } from "libs/ui/react/utils/layered";
-import { disableAnimationsOnPerformanceMode } from "libs/ui/react/utils/performance";
 
 import App from "./App.svelte";
 
@@ -14,7 +13,6 @@ const widget = Widget.getCurrent();
 await widget.init();
 
 await declareDocumentAsLayeredHitbox((e) => e.getAttribute("data-allow-mouse-events") === "true");
-disableAnimationsOnPerformanceMode();
 
 const container = getRootContainer();
 mount(App, {